requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
scikit-learn==1.7.1
numpy==2.2.6
scipy==1.15.3
joblib==1.5.1
python-dateutil==2.9.0.post0
